# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookstand API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/readers": {
            "get": {
                "tags": [
                    "Readers"
                ],
                "summary": "Get all readers",
                "operationId": "3d6b99504f2f739d7f1d2782006b663a",
                "responses": {
                    "200": {
                        "description": "List of all readers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reader"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Readers"
                ],
                "summary": "Create a new reader",
                "operationId": "3d46d28d1e1e76c11637da9985ee1e30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reader_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "reader_email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reader created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reader"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/readers/{id}": {
            "get": {
                "tags": [
                    "Readers"
                ],
                "summary": "Get a reader by ID",
                "operationId": "06d71432581ededb07b6e6ea638d43f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reader ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reader data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reader"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reader not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Readers"
                ],
                "summary": "Update a reader",
                "operationId": "14c06642d3dd6070c96096884e271a40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reader ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reader_name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "reader_email": {
                                        "type": "string",
                                        "example": "janedoe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reader updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reader"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reader not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Readers"
                ],
                "summary": "Delete a reader",
                "operationId": "bf6f1eb926b316fb6328a3f95d40975b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reader ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reader deleted successfully"
                    },
                    "404": {
                        "description": "Reader not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Reader": {
                "title": "Reader",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    ]
}